% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-databases.R
\name{download_mcdb}
\alias{download_mcdb}
\title{Download the Mammal Community Database}
\usage{
download_mcdb()
}
\value{
null
}
\description{
Saves data tables from http://esapubs.org/archive/ecol/E092/201/ to \code{analysis/data/raw_data}.
}
