% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-databases.R
\name{download_quaternarymammals}
\alias{download_quaternarymammals}
\title{Download Body Mass of Late Quaternary Mammals}
\usage{
download_quaternarymammals()
}
\value{
null
}
\description{
Saves data tables from https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/02-9003 to \code{analysis/data/raw_data}.
}
