% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-databases.R
\name{load_mcdb}
\alias{load_mcdb}
\title{Load the MCDB}
\usage{
load_mcdb()
}
\value{
list of 4 data tables: \code{species}, \code{sites}, \code{communities}, and \code{trapping}
}
\description{
Loads data tables from http://esapubs.org/archive/ecol/E092/201/ saved in \code{analysis/data/raw_data}.
}
