% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-databases.R
\name{load_quaternarymammals}
\alias{load_quaternarymammals}
\title{Load the Body Mass of Quaternary Mammals databasee}
\usage{
load_quaternarymammals()
}
\value{
list of 4 data tables: \code{species}, \code{sites}, \code{communities}, and \code{trapping}
}
\description{
Loads Body Mass of Late Quaternary Mammals
}
